Enable renovate automerge.
Automerge everything, even major updates. Usually, you define some constraints with "match..." options, but I want to test out how merging everything as long as the tests are green turns out. In theory, if a merge causes a failure, it's just a sign of an insufficient test suite (which is definitely the case here, let's find out what tests we will need in order to not break the project automatically)